moved logos into branding component
