+++ Lexicon/client/src/components/branding/Branding.js
@@ -1,7 +1,27 @@
-import React from "react"
-import "./MwLogo.css"
+import React from 'react'
+import './Branding.css'
 
-const MwLogo = props => (
+// To use logo
+    // Assign a location prop: logo__login--lex, header, footer (from Branding.css)
+export const LexLogo = props => (
+    <svg className={props.location} width="185" height="150" version="1.1" viewBox="0 0 48.948 39.688" xmlns="http://www.w3.org/2000/svg">
+        <title>Lexicon Logo</title>
+        <path id="tri_yellow" d="m34.539 0.033073c-23.026-0.022049-11.513-0.011024 0 0zm-9.9115 11.398-24.542 28.223h16.259l7.3246-10.08 7.0714-9.7307z" style={{fill:"#e4e485",paintOrder:"stroke fill markers", strokeLinecap:"round", strokeLinejoin:"round", strokeWidth:".10087"}}/>
+        <path d="m20.07 0.033577 6.5867 9.063 5.29e-4 -5.291e-4 0.8692 1.1963 6.9407 9.5514 14.396-19.81h-28.792z" style={{fill:"#d64d4d", paintOrder:"stroke fill markers", strokeLinecap:"round",strokeLinejoin:"round",strokeWidth:".050434"}}/>
+        <path id="tri_red" d="m0.085055 0.033577 17.39 19.811 7.2884-8.2253-8.4191-11.585z" style={{fill:"#d64d4d", paintOrder:"stroke fill markers", strokeLinecap:"round",strokeLinejoin:"round",strokeWidth:".050434"}}/>
+        <path className={props.color} id="tri_black-white" d="m34.467 19.844-14.396 19.811h28.792z" style={{fill:"#171717", paintOrder:"stroke fill markers", strokeLinecap:"round",strokeLinejoin:"round",strokeWidth:".050434"}}/>
+        <path id="tri_orange" d="m24.709 11.544-7.2342 8.2998h13.265z" style={{fillOpacity:".57143",fill:"#d64d4d",paintOrder:"stroke fill markers",strokeLinecap:"round",strokeLinejoin:"round",strokeWidth:".021494"}}/>
+    </svg>
+)
+
+export const LexTitle = props => (
+    <svg className={props.location} width="515" height="70" version="1.1" viewBox="0 0 136.26 18.521" xmlns="http://www.w3.org/2000/svg">
+        <title>Lexicon Title</title>
+        <path className={props.color} transform="scale(.26458)" d="m306.62 0.048828c-6.736 0-13.113 1.4683-19.133 4.4062-6.0194 2.8664-10.855 6.9153-14.51 12.146-3.6546 5.1594-5.4824 11-5.4824 17.521 0 5.9478 1.3609 11.644 4.084 17.09 2.723 5.3745 6.9508 9.8168 12.684 13.328 5.7327 3.4397 12.899 5.1602 21.498 5.1602 7.3092 0 14.154-1.3967 20.531-4.1914 0.64466-0.28708 1.0381-0.53926 1.1816-0.75391 0.21464-0.28661 0.39356-0.85883 0.53711-1.7188l1.5059-11.502c0-0.14355-0.10762-0.2505-0.32227-0.32226-0.14353-0.07176-0.32245-0.10742-0.53711-0.10742-0.28641 0-0.42968 0.10761-0.42968 0.32226-1.7915 5.7328-4.6226 9.9963-8.4922 12.791-3.798 2.7947-8.3476 4.1914-13.65 4.1914-5.6611 0-10.93-1.6489-15.803-4.9453-4.8728-3.368-8.7768-7.8819-11.715-13.543-2.938-5.7328-4.4082-11.967-4.4082-18.703 0-5.7327 1.1821-10.785 3.5469-15.156s5.6628-7.7387 9.8906-10.104c4.2996-2.4364 9.2426-3.6543 14.832-3.6543 13.974 0 22.25 5.2311 24.83 15.693 0.0718 0.21532 0.2883 0.32226 0.64648 0.32226 0.35819-0.07177 0.53711-0.21505 0.53711-0.42969l-1.1836-11.285c-0.07179-0.71684-0.21507-1.2194-0.42969-1.5059-0.14307-0.35839-0.50072-0.71603-1.0742-1.0742-2.1498-1.2182-5.0167-2.1858-8.5996-2.9023-3.5113-0.71657-7.0218-1.0742-10.533-1.0742zm82.07 0c-6.091 0-12.075 1.3252-17.951 3.9766-5.8044 2.6514-10.571 6.6288-14.297 11.932-3.7263 5.2311-5.5879 11.537-5.5879 18.918 0 6.306 1.3967 12.145 4.1914 17.52 2.8664 5.3028 6.8796 9.5306 12.039 12.684 5.2311 3.0814 11.178 4.623 17.842 4.623 6.736 0 12.97-1.5774 18.703-4.7305 5.8044-3.153 10.428-7.4523 13.867-12.898 3.4396-5.4461 5.1582-11.43 5.1582-17.951 0-5.8761-1.3967-11.429-4.1914-16.66-2.723-5.2311-6.6647-9.4232-11.824-12.576-5.1595-3.2247-11.142-4.8359-17.949-4.8359zm-388.3 1.1797c-0.14354 0-0.21484 0.21457-0.21484 0.64453s0.071299 0.64648 0.21484 0.64648c2.7947 0 4.8371 0.21457 6.127 0.64453 1.3615 0.35832 2.2575 1.0397 2.6875 2.043 0.50173 1.0032 0.75195 2.5787 0.75195 4.7285v49.768c0 2.1498-0.25022 3.763-0.75195 4.8379-0.42989 1.0033-1.3259 1.6828-2.6875 2.041-1.2899 0.35819-3.3322 0.53711-6.127 0.53711-0.14354 0-0.21484 0.21457-0.21484 0.64453 0 0.42996 0.071299 0.64648 0.21484 0.64648h49.445c0.64526 0 1.0744-0.10761 1.2891-0.32226 0.21465-0.28641 0.32226-0.71751 0.32226-1.291 0-3.6546 0.3222-8.3116 0.9668-13.973 0-0.21463-0.21457-0.32227-0.64453-0.32227-0.35819-0.07108-0.53711 2.17e-4 -0.53711 0.21484-1.2899 9.1008-5.4819 13.652-12.576 13.652h-13.758c-2.7231 0-4.7297-0.17892-6.0195-0.53711-1.2181-0.35832-2.0783-1.004-2.5801-1.9355-0.42996-1.0033-0.64453-2.543-0.64453-4.6211v-49.232c0-2.1498 0.21457-3.7253 0.64453-4.7285 0.50166-1.075 1.3976-1.7922 2.6875-2.1504 1.3615-0.42996 3.4396-0.64453 6.2344-0.64453 0.14354 0 0.21484-0.21653 0.21484-0.64648s-0.0713-0.64453-0.21484-0.64453l-5.0508 0.10742c-3.0097 0.14354-5.4832 0.2168-7.418 0.2168-1.7198 0-4.0841-0.073252-7.0938-0.2168zm69.277 0c-0.14355 0-0.21484 0.21457-0.21484 0.64453s0.0713 0.64648 0.21484 0.64648c2.7947 1e-7 4.8371 0.21457 6.127 0.64453 1.3615 0.35832 2.2575 1.0397 2.6875 2.043 0.50173 1.0032 0.75195 2.5787 0.75195 4.7285v49.768c0 2.1498-0.25022 3.763-0.75195 4.8379-0.42996 1.0033-1.326 1.6828-2.6875 2.041-1.2899 0.35819-3.3322 0.53711-6.127 0.53711-0.14355 0-0.21484 0.21653-0.21484 0.64648 0 0.42996 0.0713 0.64453 0.21484 0.64453h44.393c0.64529 0 1.0744-0.10761 1.2891-0.32226 0.28642-0.28641 0.42969-0.71751 0.42969-1.291 0-4.8012 0.3222-9.4581 0.9668-13.973 0-0.21465-0.21458-0.32227-0.64453-0.32227-0.42996-0.07111-0.64454 1.92e-4 -0.64454 0.21484l2e-3 -2e-3c-0.57324 4.0846-1.9004 7.3825-3.9785 9.8906-2.0064 2.508-4.4422 3.7617-7.3086 3.7617h-9.8887c-2.7947 0-4.8371-0.17892-6.127-0.53711-1.218-0.35832-2.0783-1.004-2.5801-1.9355-0.42996-1.0033-0.64453-2.543-0.64453-4.6211v-24.83h13.865c3.4397 0 5.8416 0.64372 7.2031 1.9336s2.041 3.5468 2.041 6.7715c0 0.14355 0.21457 0.2168 0.64453 0.2168s0.64648-0.07325 0.64648-0.2168l-0.10937-5.0508c-0.0718-1.0033-0.10742-2.5449-0.10742-4.623l0.10742-8.4902c0-0.14354-0.21457-0.2168-0.64453-0.2168s-0.64453 0.07325-0.64453 0.2168c0 2.7231-0.64567 4.6562-1.9356 5.8027-1.2899 1.0749-3.5826 1.6133-6.8789 1.6133h-14.188v-21.82c0-2.1498 0.21457-3.7273 0.64453-4.7305 0.50159-1.0034 1.3976-1.6828 2.6875-2.041 1.2899-0.35819 3.2965-0.53711 6.0195-0.53711h7.416c3.153 0 5.6961 1.0748 7.6309 3.2246 1.9348 2.1497 2.9024 5.0524 2.9024 8.707 0 0.21464 0.18087 0.32031 0.53906 0.32031 0.42996 0 0.64453-0.10567 0.64453-0.32031l-0.2168-13.008c0-0.6446-0.3222-0.9668-0.96679-0.9668zm426.14 0c-0.14362 0-0.21484 0.21653-0.21484 0.64648s0.0712 0.64453 0.21484 0.64453c3.0097 0 5.0878 0.75101 6.2344 2.2559 1.1466 1.5049 1.7207 4.0856 1.7207 7.7402v47.08l-42.459-52.777c-0.93155-1.1465-1.7918-2.3644-2.5801-3.6543-0.43001-0.78855-0.82344-1.2893-1.1816-1.5039-0.28654-0.28641-0.67992-0.42969-1.1816-0.42969-0.42996 0-1.0399 0.035649-1.8281 0.10742-0.78825 0.071773-1.7915 0.10742-3.0098 0.10742-1.4332 0-2.5438-0.035649-3.332-0.10742l-3.8691-0.10742c-0.21468-1e-7 -0.32227 0.21457-0.32227 0.64453s0.10759 0.64453 0.32227 0.64453c2.6514 0 5.2302 1.0748 7.7383 3.2246v51.486c0 3.7263-0.57222 6.307-1.7188 7.7402-1.1466 1.4332-3.2604 2.1504-6.3418 2.1504-0.14362 0-0.2168 0.21457-0.2168 0.64453 0 0.42996 0.0732 0.64453 0.2168 0.64453 1.8632 0 3.2598-0.03566 4.1914-0.10742l5.2676-0.10742 5.1582 0.10742c0.93155 0.07176 2.2587 0.10742 3.9785 0.10742 0.21468 0 0.32227-0.21457 0.32227-0.64453 0-0.42996-0.10759-0.64453-0.32227-0.64453-3.0097 0-5.1236-0.71717-6.3418-2.1504-1.2182-1.5049-1.8281-4.0856-1.8281-7.7402v-49.229l1.8281 2.041 48.584 59.766c0.14355 0.14352 0.39571 0.17918 0.75391 0.10742 0.35817-0.07108 0.53711-0.17872 0.53711-0.32227v-57.076c0-3.6546 0.60793-6.2354 1.8262-7.7402 1.2899-1.5049 3.4753-2.2578 6.5566-2.2578 0.21543 0 0.32422-0.21457 0.32422-0.64453s-0.10879-0.64453-0.32422-0.64453l-4.1914 0.10547c-2.2931 0.14354-4.0494 0.2168-5.2676 0.2168-1.2899 0-3.0084-0.073252-5.1582-0.2168zm-359.66 0.00195c-0.14355 0-0.21484 0.21457-0.21484 0.64453s0.0713 0.64648 0.21484 0.64648c2.5798 0 4.8029 0.57215 6.666 1.7188 1.8632 1.1465 3.7248 3.1174 5.5879 5.9121l17.307 26.119-16.23 22.25c-2.0782 2.7947-4.2635 4.9463-6.5566 6.4512-2.2931 1.4332-4.6593 2.1484-7.0957 2.1484-0.21466 0-0.32032 0.21457-0.32032 0.64453 0 0.42996 0.10566 0.64648 0.32032 0.64648 1.7915 0 3.0828-0.03566 3.8711-0.10742l4.4062-0.10938 7.2012 0.10938c1.0749 0.07176 2.6881 0.10742 4.8379 0.10742 0.21465 0 0.32227-0.21653 0.32227-0.64648 0-0.42996-0.10762-0.64453-0.32227-0.64453-4.3712 0-6.5566-1.1821-6.5566-3.5469 0-1.2899 0.71529-2.9746 2.1484-5.0527l14.941-20.852 16.662 25.26c0.6446 0.93135 0.9668 1.6843 0.9668 2.2578 0 0.71678-0.46688 1.2175-1.3984 1.5039-0.93155 0.28641-2.4713 0.42969-4.6211 0.42969-0.14355 0-0.2168 0.21457-0.2168 0.64453 0 0.42996 0.0733 0.64453 0.2168 0.64453 2.2214 0 3.9758-0.03566 5.2656-0.10742l8.0625-0.10742 5.6973 0.10742c1.0749 0.07176 2.5431 0.10742 4.4062 0.10742 0.14355 0 0.21484-0.21457 0.21484-0.64453 0-0.42996-0.0713-0.64453-0.21484-0.64453-3.153 0-5.7318-0.53643-7.7383-1.6113-1.9348-1.1466-3.8342-3.1532-5.6973-6.0195l-17.844-26.766 15.049-20.529c4.7295-6.4494 9.5315-9.6758 14.404-9.6758 0.21463 1e-7 0.32031-0.21457 0.32031-0.64453s-0.10568-0.64453-0.32031-0.64453l-3.8711 0.10742c-1.7198 0.14354-3.188 0.21484-4.4062 0.21484-2.0782 0-4.4782-0.071299-7.2012-0.21484l-4.8379-0.10742c-0.2153-1e-7 -0.32226 0.21457-0.32226 0.64453s0.10696 0.64453 0.32226 0.64453c4.1562 0 6.2344 1.2198 6.2344 3.6562 0 1.4331-0.86025 3.404-2.5801 5.9121l-13.65 19.24-16.338-24.508c-0.6446-1.0031-0.9668-1.7561-0.9668-2.2578 0-0.71684 0.46493-1.2175 1.3965-1.5039 0.93162-0.35819 2.4732-0.53711 4.623-0.53711 0.21463 0 0.32032-0.21653 0.32032-0.64648-1e-5 -0.42996-0.10569-0.64453-0.32032-0.64453l-5.5898 0.10742c-3.4396 0.14354-6.3065 0.21484-8.5996 0.21484l-3.8691-0.10742c-1.8631-0.14354-3.226-0.21484-4.0859-0.21484zm84.326 0c-0.21465 0-0.32226 0.21457-0.32226 0.64453s0.10761 0.64453 0.32226 0.64453c2.7947 0 4.8371 0.21653 6.127 0.64648 1.3614 0.35832 2.2932 1.0378 2.7949 2.041 0.50173 1.0032 0.75195 2.5807 0.75195 4.7305v49.768c0 2.1498-0.25022 3.761-0.75195 4.8359-0.50166 1.0033-1.4334 1.6848-2.7949 2.043-1.2899 0.35819-3.3322 0.53711-6.127 0.53711-0.21465 0-0.32226 0.21457-0.32226 0.64453 0 0.42996 0.10761 0.64453 0.32226 0.64453 2.2215 0 3.9758-0.03566 5.2656-0.10742l7.2031-0.10742 7.5234 0.10742c1.2182 0.07176 2.9029 0.10742 5.0527 0.10742 0.14355 0 0.21485-0.21457 0.21485-0.64453 0-0.42996-0.0713-0.64453-0.21485-0.64453-2.8664 0-4.9445-0.17892-6.2344-0.53711-1.2897-0.35832-2.1858-1.0398-2.6875-2.043-0.42996-1.0033-0.64453-2.6145-0.64453-4.8359v-49.553c0-2.1498 0.21457-3.7272 0.64453-4.7305 0.5016-1.0749 1.3976-1.8259 2.6875-2.2559 1.3615-0.42996 3.4396-0.64648 6.2344-0.64648 0.14355 0 0.21485-0.21457 0.21485-0.64453s-0.0713-0.64453-0.21485-0.64453l-5.0527 0.10742c-3.0097 0.14354-5.517 0.21484-7.5234 0.21484-1.7915 0-4.1934-0.071299-7.2031-0.21484zm163.71 1.1816c6.306 0 11.823 1.4702 16.553 4.4082 4.8012 2.9381 8.4925 7.0586 11.072 12.361 2.6514 5.2312 3.9766 11.213 3.9766 17.949 0 6.3777-1.1821 11.861-3.5469 16.447-2.3648 4.5862-5.5892 8.061-9.6738 10.426-4.0845 2.3648-8.6699 3.5469-13.758 3.5469-6.0193 0-11.394-1.6113-16.123-4.8359-4.7295-3.2247-8.4209-7.5975-11.072-13.115-2.5798-5.5894-3.8691-11.715-3.8691-18.379 0-8.8858 2.3285-15.909 6.9863-21.068 4.6579-5.1595 11.143-7.7402 19.455-7.7402z" style={{fill:"#171717",whiteSpace:"pre"}}/>
+    </svg>
+)
+
+export const MwLogo = props => (
     <svg className={props.location} width="125" height="125" version="1.1" viewBox="0 0 33.073 33.073" xmlns="http://www.w3.org/2000/svg">
         <title>Merriam Webster Logo</title>
         <path id="background" d="m0.1737 16.537c0-9.05 7.3519-16.386 16.422-16.386 9.0696 0 16.305 7.3362 16.305 16.386 0 9.05-7.2351 16.386-16.305 16.386-9.0696 0-16.422-7.3362-16.422-16.386" style={{fill:"#fff"}}/>
@@ -10,6 +30,4 @@ const MwLogo = props => (
         <path id="trademark" d="m27.908 31.109c0.11756 0 0.22253-8e-3 0.22253-0.14905 0-0.11337-0.10287-0.13436-0.19944-0.13436h-0.18894v0.28341zm-0.16585 0.51399h-0.13821v-0.91497h0.34815c0.21623 0 0.3233 0.0798 0.3233 0.26032 0 0.1634-0.10287 0.23513-0.23723 0.25158l0.26032 0.40307h-0.15535l-0.24143-0.39643h-0.15955zm0.1683 0.20364c0.35864 0 0.64205-0.28131 0.64205-0.6634 0-0.37543-0.28341-0.65884-0.64205-0.65884-0.36319 0-0.6466 0.28341-0.6466 0.65884 0 0.38209 0.28341 0.6634 0.6466 0.6634m-0.80615-0.6634c0-0.45521 0.36948-0.79075 0.80615-0.79075 0.43177 0 0.80125 0.33554 0.80125 0.79075 0 0.45976-0.36948 0.79566-0.80125 0.79566-0.43667 0-0.80615-0.3359-0.80615-0.79566" style={{fill:"#d91920"}}/>
         <path id="text" d="m28.356 14.782h1.2859v-0.47236h-1.2859zm-4.5633 4.9027c0.52729 1e-3 0.75577 0.37684 0.75717 0.86704 0 0.0273-1e-3 0.0549-2e-3 0.0847l-1.5997-0.0144c0.0406-0.46605 0.35024-0.93841 0.84498-0.93736m1.3765 2.4321c-0.27151 0.28657-0.53918 0.52589-0.92616 0.52589-0.0196 0-0.0392-7e-4 -0.0581-2e-3 -1.099-0.0665-1.2505-0.89537-1.2523-1.8404l1.9318 0.0143c0.20888-7e-4 0.29426-0.0269 0.29426-0.303-1e-3 -0.61546-0.52869-1.0444-1.261-1.0462-0.86808 2e-3 -1.5661 0.65045-1.5675 1.803 1e-3 1.1739 0.69209 1.8226 1.5675 1.8243 0.35864-7e-4 0.60847-0.0486 0.66515-0.0812l0.84219-0.87543zm-12.595-2.409c0.52694 1e-3 0.75507 0.37754 0.75647 0.86774 0 0.0273-7e-4 0.0553-1e-3 0.084l-1.6001-0.0136c0.0406-0.46606 0.34989-0.93807 0.84499-0.93807m0.44961 2.9584c-0.0192 0-0.0388-7e-4 -0.0574-1e-3 -1.0987-0.0661-1.2509-0.89573-1.2526-1.8408l1.9318 0.0143c0.20924-1e-3 0.29391-0.0266 0.29391-0.30266-7e-4 -0.61546-0.52763-1.0451-1.2607-1.0465-0.86843 2e-3 -1.5654 0.65045-1.5675 1.803 0.0021 1.1739 0.69174 1.8229 1.5675 1.8243 0.35934-7e-4 0.60881-0.049 0.6648-0.0812l0.84254-0.87578-0.23583-0.0189c-0.27187 0.28586-0.53918 0.52554-0.92652 0.52554m13.94 0.30895h-0.0206zm-1.8264-0.015v-2e-3zm1.3446-2.7477-0.0976 0.1571v-0.84674l-0.1704-0.0637-5e-3 0.0137-6e-3 -0.0137c-0.28377 0.12876-0.70049 0.24353-1.0332 0.28586l-0.0129 0.17705c0.21518 0.0143 0.42896 0.05 0.61406 0.14205l-8e-3 2.6004-0.62876 0.10706-0.0115 0.20364 1.8551 0.0143 0.0143-0.20364-0.6375-0.13576 0.0115-1.7635c7e-4 -0.12036 0.33029-0.93526 0.71623-0.93596 0.084-7e-4 0.20608 0.0626 0.28061 0.11861l0.18929 0.0182 0.13681-0.48215c-0.12386-0.10637-0.41287-0.12352-0.56053-0.12422-0.38383 2e-3 -0.49405 0.44087-0.64695 0.73128m-8.5367-7.2487c-0.11896-0.10217-0.39362-0.11861-0.53463-0.11931-0.36564 2e-3 -0.47061 0.42057-0.61581 0.69594l-0.0924 0.1473 7e-4 -0.80265-0.16235-0.0619-4e-3 0.014-6e-3 -0.0133c-0.26977 0.12211-0.6669 0.23163-0.98285 0.27221l-0.0122 0.169c0.20468 0.0143 0.40797 0.0486 0.58362 0.13611l-8e-3 2.4727-0.59727 0.10182-0.0122 0.19384h0.0147v3.5e-4h0.0455l1.7061 0.0147 0.0144-0.1949-0.60742-0.13016 0.0122-1.6767c0-0.11442 0.31386-0.88908 0.6802-0.88978 0.0794-7e-4 0.19559 0.0591 0.26626 0.11232l0.18125 0.0178zm-3.1924 3.2061v3.5e-4l-0.0196-3.5e-4zm-1.1424-0.29601-0.59726 0.10147-0.0122 0.19489 1.7663 0.0143 0.0143-0.1949-0.60742-0.13016 0.0115-1.6767c7e-4 -0.11442 0.31561-0.88943 0.68089-0.89013 0.0794-3.5e-4 0.19524 0.0595 0.26627 0.11302l0.1809 0.0175 0.13121-0.45941c-0.11897-0.10147-0.39363-0.11861-0.53534-0.11966-0.36494 2e-3 -0.46991 0.42057-0.61511 0.69629l-0.0917 0.1473v-0.80265l-0.16235-0.0616-4e-3 0.0136-7e-3 -0.0136c-0.26977 0.12246-0.6669 0.23198-0.98285 0.27256l-0.0123 0.16865c0.20469 0.0144 0.40797 0.049 0.58432 0.13611zm-1.9559-2.8422c0.49859 7e-4 0.71413 0.35584 0.71553 0.8198 0 0.0259-7e-4 0.0532-1e-3 0.0801l-1.5133-0.0133c0.03814-0.44156 0.331-0.88733 0.79916-0.88663m0.0987 3.2274c0.34079 0 0.57662-0.0458 0.6305-0.0766l0.79881-0.83065-0.22463-0.0185c-0.25822 0.27081-0.51155 0.49789-0.87823 0.49789-0.0178 0-0.036 0-0.0539-2e-3 -1.0399-0.0626-1.1833-0.84674-1.1851-1.7421h0.02169l1.8068 0.014c0.19804 0 0.27957-0.0259 0.27957-0.28761-1e-3 -0.58327-0.5007-0.99019-1.1956-0.99089-0.82259 1e-3 -1.4828 0.61581-1.4849 1.7085 0.0021 1.1116 0.65605 1.726 1.4849 1.7278m-5.9555-0.0889-0.01749-3.5e-4h0.01749zm4.796-3.5e-4v3.5e-4l-0.02169-3.5e-4zm-1.8999-0.0147v-2e-3zm-2.8674-0.16724-0.62771-0.13051 0.34954-3.7771 0.03254-0.0931 1.6833 4.0605 0.09307 9e-3 1.6368-4.0262 0.04514 0.0829 0.3226 3.7592-0.67039 0.10112-0.0119 0.19664 1.9286 0.0143 0.01434-0.19664-0.58607-0.13051-0.41917-4.1721 0.65675-0.10007-0.0021-0.0143 0.01435-1e-3 -0.0077-0.16725-1.4325-0.0136-1.4307 3.4793-1.451-3.4709-1.4573-8e-3 -0.02204 0.18089 0.71378 0.13156-0.40098 4.1693-0.55248 0.10322-0.0112 0.19664 1.578 0.0143zm12.176-2.7372h-3.5e-4l3.5e-4 -2e-3zm-0.0714 2.9044h-3.5e-4l3.5e-4 -5e-3zm0.62106-0.25612-0.62351 0.10147-0.0122 0.1697 1.7876 0.0143v-0.0143h0.0143v-0.15535l-0.62211-0.13016 0.0112-2.9996-0.13296-0.0616c-0.27536 0.12246-0.66444 0.23198-0.97445 0.27256l-0.0122 0.16865c0.20119 0.0144 0.39993 0.049 0.57243 0.13611zm2.9926-1.3191v0.22498l-0.0147 0.74912c-0.20644 0.12841-0.52694 0.29496-0.71868 0.29426-0.25367-7e-4 -0.51784-0.1795-0.51854-0.41427-2e-3 -0.46186 0.49265-0.59342 1.2519-0.85409m0-0.0147v4e-3c-0.0623 0.0196-0.12036 0.0381-0.17879 0.0567 0.0588-0.0199 0.11651-0.0395 0.17879-0.0609m-0.0252 1.627c-3.5e-4 0-4e-3 7e-4 -4e-3 7e-4l8e-3 -0.10112zm-1.0843 9e-3h0.0157c0.41602-0.0105 0.75262-0.16725 1.0707-0.41672l-0.0175 0.42162c7e-4 7e-4 2e-3 1e-3 6e-3 1e-3 2e-3 0 7e-3 -1e-3 9e-3 -1e-3v1e-3s9e-3 -2e-3 0.013-3e-3c0.12841-0.0147 0.79705-0.16585 1.0987-0.22218l0.021-4e-3 -0.10357-0.16095-0.4881-0.0507 7e-4 -2.0899c-7e-4 -0.60426-0.26207-0.87858-0.68474-0.87928-0.0196 0-0.0399-7e-4 -0.0605-7e-4 -0.0602 0-0.12386 4e-3 -0.18019 0.0367l-0.76592 0.45766c-0.0777 0.0504-0.35164 0.19979-0.36389 0.33275l-0.0259 0.37998 0.0525 0.0406 0.39958-0.0801c0.0497-0.0112 0.12176-8e-3 0.12176-0.0913v-0.0115c-4e-3 -0.0658-7e-3 -0.12946-7e-3 -0.18964 2e-3 -0.26627 0.0647-0.46501 0.54619-0.53359 0.0178-3e-3 0.035-5e-3 0.0521-5e-3 0.24177 2e-3 0.38488 0.25612 0.38558 0.74842v0.50455c-0.32855 0.10217-0.68859 0.24352-1.0731 0.37193-0.29111 0.0941-0.74247 0.23303-0.74247 0.81035 1e-3 0.27257 0.31245 0.63121 0.72112 0.63191m-1.6557-4.2214c0.16025 0 0.31281-0.12912 0.31351-0.30441-7e-4 -0.1795-0.13366-0.34045-0.31351-0.34115-0.17459 1e-3 -0.3009 0.1746-0.3009 0.34115 7e-4 0.1634 0.1459 0.30441 0.3009 0.30441m9.5678 4.1655h-0.0179zm-1.9752 0h-0.0182zm-2.001 0h-0.0182zm-1.605-0.0151h-3.5e-4l3.5e-4 -4e-3zm-0.0154-2.9167h-7e-4l7e-4 -4e-3zm5.1109 0.69628c2e-3 -0.68194-0.34255-1.0675-1.0665-1.0868-3e-3 -7e-4 -7e-3 -7e-4 -0.0105-7e-4 -0.0878 0-0.22778 0.035-0.42442 0.18824l-0.50665 0.43317c-0.18159-0.3443-0.52729-0.61337-0.93456-0.61442-0.26732-7e-4 -0.36494 0.0367-0.57417 0.20119l-0.48461 0.31351v-0.45871l-0.15569-0.0616c-0.26977 0.12246-0.64975 0.23198-0.95311 0.27256l-0.0157 0.13086c0.19699 0.015 0.39118 0.0493 0.56053 0.13681l-5e-3 2.5017-0.52869 0.10217-0.0115 0.1774 1.634 0.0143 0.0144-0.16095-0.55073-0.1473 0.0112-2.2421c0.24003-0.1837 0.514-0.40343 0.80301-0.40273 0.61301 7e-4 0.62806 0.62036 0.62876 1.1676v1.4916l-0.52869 0.11862-0.0109 0.16095 1.634 0.0143 0.0143-0.16095-0.55073-0.1473 0.0112-2.2491c0.23408-0.1725 0.54863-0.35304 0.84149-0.35304 0.55143 2e-3 0.58957 0.46675 0.59062 0.9587 0 0.0833-3.5e-4 0.1662-3.5e-4 0.24843v1.4094l-0.52869 0.11862-0.0115 0.16095 1.6095 0.0143 0.0143-0.16095-0.52589-0.1473zm-5.5916 8.3971-0.0196 0.0189c-0.1578 0.15815-0.40133 0.40273-0.59727 0.40133-0.37369-8e-3 -0.51924-0.26522-0.52134-0.73408 0-0.021 0-0.0416 1e-3 -0.0619v-1.8698h0.0241l0.99475 0.014 0.0144-0.34779-1.0476-0.0144 0.0143-0.89082-0.1795-0.0144c-0.1431 0.5-0.53918 0.79636-0.92476 1.0756l-7e-3 0.1774h0.0151v0.01435h0.49579l-0.0143 2.142c7e-4 0.6704 0.43247 0.91077 0.81945 0.91217 0.0151 0 0.0294-7e-4 0.0448-7e-4 0.54758-0.0273 0.92477-0.57242 0.92547-0.57312l-0.0336-0.22218zm-11.215-3.9632-0.04059-0.0108 0.04094 0.0101zm-0.63925-0.33345v-3.4e-4l0.01959 3.4e-4zm-4.5364 0.36144h6.99e-4l0.04444 0.14066zm0.59202-0.33064h7e-4l-0.0014 8e-3zm5.6186 0.15185 0.0109-0.18299-1.6994-0.014-0.01435 0.18334 0.6634 0.1774-0.09902 0.28306-1.1127 3.1438c-0.12106-0.36319-1.0332-3.1102-1.0955-3.3432l0.08537-0.13331 0.55423-0.0826 0.01225-0.20014-1.838-0.0119 7e-4 0.0144h-0.01505v0.18544l0.5035 0.11547 0.09167 0.16725c-0.06333 0.23582-1.0161 2.9408-1.1441 3.3044l-1.1008-3.4314 0.03254-9e-3 0.53569-0.13261 0.02519-0.19909-0.014-3.5e-4v-3.5e-4l-0.06753-3.5e-4 -1.831-0.0185-0.01784 0.18719 0.65395 0.162 1.5087 4.7645 0.23093 0.0105 1.374-3.7907 1.2743 3.7666 0.24527 0.0115 1.6658-4.7715zm4.0332 1.3058c-0.11336 0-0.21798 0.0213-0.32785 0.0357l-0.58362 0.37193 0.0847 0.0707c0.1578-0.0479 0.31561-0.0759 0.46606-0.0759 0.81315 1e-3 1.1739 0.75577 1.1746 1.521-7e-4 0.7939-0.34779 1.4657-1.0633 1.4671-0.47061-7e-4 -0.75087-0.3247-0.75157-0.79286v-4.6571l-0.14416-0.0644c-0.29356 0.12981-0.64975 0.24388-0.9804 0.28691l-8e-3 0.13331c0.21344 0.0136 0.36634 0.0486 0.55213 0.14101l-0.013 3.7155c0 0.16795 1e-3 0.33345 1e-3 0.49999 0 0.29601-4e-3 0.59342-0.0259 0.90553l0.13261 0.0717 0.26872-0.25997c0.22218 0.13366 0.50315 0.23898 0.75962 0.23968 1.2687-2e-3 1.8187-0.91252 1.8201-1.9006-1e-3 -1.1162-0.5028-1.7075-1.3625-1.7092m1.6032 2.3852-3.5e-4 -1e-3 0.0206 1e-3zm0.53534-1.5784c0-0.38803 0.28026-0.62736 0.60566-0.62806 0.44787 7e-4 0.66375 0.345 0.66515 0.80021l0.24283 0.0143 0.0273-0.69978h-0.0144l9e-3 -0.0112c-0.27747-0.24737-0.58887-0.31595-0.94297-0.31595-0.54863 1e-3 -1.0598 0.37928-1.0605 1.0189 8e-3 1.3303 1.7323 0.76101 1.7173 1.6644-1e-3 0.41812-0.26102 0.71343-0.66899 0.71448h-0.0182c-0.51189-0.0102-0.8303-0.43177-0.84394-0.9797l-0.26872-0.0136 1e-3 0.0143h-1e-3l6e-3 0.0682 0.0703 0.92336c0.3114 0.1781 0.6536 0.22463 0.98775 0.22463 0.6382-1e-3 1.2726-0.35269 1.2733-1.1228-7e-3 -1.2785-1.8034-0.90587-1.7876-1.6718" style={{fill:"#1b458f"}}/>
     </svg>
-    )
-
-export default MwLogo
+)
