--- Lexicon/client/src/components/branding/LexLogo.css
@@ -1,24 +1,0 @@
-.logo__login--lex {
-    width: 6em;
-    height: 6em;
-}
-
-.logo__header--lex {
-    width: 1.25em;
-    height: 1.25em;
-    margin-left: 0.85em;
-}
-
-.logo__footer--lex {
-    width: 2em;
-    height: 2em;
-    margin-bottom: 0.5em;
-}
-
-.logo__white {
-    fill: #FCFCFC !important;
-}
-
-.logo__black {
-    fill: var(--black) !important;
-}
