--- Lexicon/client/src/components/branding/LexTitle.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import "./LexTitle.css"
-
-const LexTitle = props => (
-    <svg className={props.location} width="515" height="70" version="1.1" viewBox="0 0 136.26 18.521" xmlns="http://www.w3.org/2000/svg">
-        <title>Lexicon Title</title>
-        <path className={props.color} transform="scale(.26458)" d="m306.62 0.048828c-6.736 0-13.113 1.4683-19.133 4.4062-6.0194 2.8664-10.855 6.9153-14.51 12.146-3.6546 5.1594-5.4824 11-5.4824 17.521 0 5.9478 1.3609 11.644 4.084 17.09 2.723 5.3745 6.9508 9.8168 12.684 13.328 5.7327 3.4397 12.899 5.1602 21.498 5.1602 7.3092 0 14.154-1.3967 20.531-4.1914 0.64466-0.28708 1.0381-0.53926 1.1816-0.75391 0.21464-0.28661 0.39356-0.85883 0.53711-1.7188l1.5059-11.502c0-0.14355-0.10762-0.2505-0.32227-0.32226-0.14353-0.07176-0.32245-0.10742-0.53711-0.10742-0.28641 0-0.42968 0.10761-0.42968 0.32226-1.7915 5.7328-4.6226 9.9963-8.4922 12.791-3.798 2.7947-8.3476 4.1914-13.65 4.1914-5.6611 0-10.93-1.6489-15.803-4.9453-4.8728-3.368-8.7768-7.8819-11.715-13.543-2.938-5.7328-4.4082-11.967-4.4082-18.703 0-5.7327 1.1821-10.785 3.5469-15.156s5.6628-7.7387 9.8906-10.104c4.2996-2.4364 9.2426-3.6543 14.832-3.6543 13.974 0 22.25 5.2311 24.83 15.693 0.0718 0.21532 0.2883 0.32226 0.64648 0.32226 0.35819-0.07177 0.53711-0.21505 0.53711-0.42969l-1.1836-11.285c-0.07179-0.71684-0.21507-1.2194-0.42969-1.5059-0.14307-0.35839-0.50072-0.71603-1.0742-1.0742-2.1498-1.2182-5.0167-2.1858-8.5996-2.9023-3.5113-0.71657-7.0218-1.0742-10.533-1.0742zm82.07 0c-6.091 0-12.075 1.3252-17.951 3.9766-5.8044 2.6514-10.571 6.6288-14.297 11.932-3.7263 5.2311-5.5879 11.537-5.5879 18.918 0 6.306 1.3967 12.145 4.1914 17.52 2.8664 5.3028 6.8796 9.5306 12.039 12.684 5.2311 3.0814 11.178 4.623 17.842 4.623 6.736 0 12.97-1.5774 18.703-4.7305 5.8044-3.153 10.428-7.4523 13.867-12.898 3.4396-5.4461 5.1582-11.43 5.1582-17.951 0-5.8761-1.3967-11.429-4.1914-16.66-2.723-5.2311-6.6647-9.4232-11.824-12.576-5.1595-3.2247-11.142-4.8359-17.949-4.8359zm-388.3 1.1797c-0.14354 0-0.21484 0.21457-0.21484 0.64453s0.071299 0.64648 0.21484 0.64648c2.7947 0 4.8371 0.21457 6.127 0.64453 1.3615 0.35832 2.2575 1.0397 2.6875 2.043 0.50173 1.0032 0.75195 2.5787 0.75195 4.7285v49.768c0 2.1498-0.25022 3.763-0.75195 4.8379-0.42989 1.0033-1.3259 1.6828-2.6875 2.041-1.2899 0.35819-3.3322 0.53711-6.127 0.53711-0.14354 0-0.21484 0.21457-0.21484 0.64453 0 0.42996 0.071299 0.64648 0.21484 0.64648h49.445c0.64526 0 1.0744-0.10761 1.2891-0.32226 0.21465-0.28641 0.32226-0.71751 0.32226-1.291 0-3.6546 0.3222-8.3116 0.9668-13.973 0-0.21463-0.21457-0.32227-0.64453-0.32227-0.35819-0.07108-0.53711 2.17e-4 -0.53711 0.21484-1.2899 9.1008-5.4819 13.652-12.576 13.652h-13.758c-2.7231 0-4.7297-0.17892-6.0195-0.53711-1.2181-0.35832-2.0783-1.004-2.5801-1.9355-0.42996-1.0033-0.64453-2.543-0.64453-4.6211v-49.232c0-2.1498 0.21457-3.7253 0.64453-4.7285 0.50166-1.075 1.3976-1.7922 2.6875-2.1504 1.3615-0.42996 3.4396-0.64453 6.2344-0.64453 0.14354 0 0.21484-0.21653 0.21484-0.64648s-0.0713-0.64453-0.21484-0.64453l-5.0508 0.10742c-3.0097 0.14354-5.4832 0.2168-7.418 0.2168-1.7198 0-4.0841-0.073252-7.0938-0.2168zm69.277 0c-0.14355 0-0.21484 0.21457-0.21484 0.64453s0.0713 0.64648 0.21484 0.64648c2.7947 1e-7 4.8371 0.21457 6.127 0.64453 1.3615 0.35832 2.2575 1.0397 2.6875 2.043 0.50173 1.0032 0.75195 2.5787 0.75195 4.7285v49.768c0 2.1498-0.25022 3.763-0.75195 4.8379-0.42996 1.0033-1.326 1.6828-2.6875 2.041-1.2899 0.35819-3.3322 0.53711-6.127 0.53711-0.14355 0-0.21484 0.21653-0.21484 0.64648 0 0.42996 0.0713 0.64453 0.21484 0.64453h44.393c0.64529 0 1.0744-0.10761 1.2891-0.32226 0.28642-0.28641 0.42969-0.71751 0.42969-1.291 0-4.8012 0.3222-9.4581 0.9668-13.973 0-0.21465-0.21458-0.32227-0.64453-0.32227-0.42996-0.07111-0.64454 1.92e-4 -0.64454 0.21484l2e-3 -2e-3c-0.57324 4.0846-1.9004 7.3825-3.9785 9.8906-2.0064 2.508-4.4422 3.7617-7.3086 3.7617h-9.8887c-2.7947 0-4.8371-0.17892-6.127-0.53711-1.218-0.35832-2.0783-1.004-2.5801-1.9355-0.42996-1.0033-0.64453-2.543-0.64453-4.6211v-24.83h13.865c3.4397 0 5.8416 0.64372 7.2031 1.9336s2.041 3.5468 2.041 6.7715c0 0.14355 0.21457 0.2168 0.64453 0.2168s0.64648-0.07325 0.64648-0.2168l-0.10937-5.0508c-0.0718-1.0033-0.10742-2.5449-0.10742-4.623l0.10742-8.4902c0-0.14354-0.21457-0.2168-0.64453-0.2168s-0.64453 0.07325-0.64453 0.2168c0 2.7231-0.64567 4.6562-1.9356 5.8027-1.2899 1.0749-3.5826 1.6133-6.8789 1.6133h-14.188v-21.82c0-2.1498 0.21457-3.7273 0.64453-4.7305 0.50159-1.0034 1.3976-1.6828 2.6875-2.041 1.2899-0.35819 3.2965-0.53711 6.0195-0.53711h7.416c3.153 0 5.6961 1.0748 7.6309 3.2246 1.9348 2.1497 2.9024 5.0524 2.9024 8.707 0 0.21464 0.18087 0.32031 0.53906 0.32031 0.42996 0 0.64453-0.10567 0.64453-0.32031l-0.2168-13.008c0-0.6446-0.3222-0.9668-0.96679-0.9668zm426.14 0c-0.14362 0-0.21484 0.21653-0.21484 0.64648s0.0712 0.64453 0.21484 0.64453c3.0097 0 5.0878 0.75101 6.2344 2.2559 1.1466 1.5049 1.7207 4.0856 1.7207 7.7402v47.08l-42.459-52.777c-0.93155-1.1465-1.7918-2.3644-2.5801-3.6543-0.43001-0.78855-0.82344-1.2893-1.1816-1.5039-0.28654-0.28641-0.67992-0.42969-1.1816-0.42969-0.42996 0-1.0399 0.035649-1.8281 0.10742-0.78825 0.071773-1.7915 0.10742-3.0098 0.10742-1.4332 0-2.5438-0.035649-3.332-0.10742l-3.8691-0.10742c-0.21468-1e-7 -0.32227 0.21457-0.32227 0.64453s0.10759 0.64453 0.32227 0.64453c2.6514 0 5.2302 1.0748 7.7383 3.2246v51.486c0 3.7263-0.57222 6.307-1.7188 7.7402-1.1466 1.4332-3.2604 2.1504-6.3418 2.1504-0.14362 0-0.2168 0.21457-0.2168 0.64453 0 0.42996 0.0732 0.64453 0.2168 0.64453 1.8632 0 3.2598-0.03566 4.1914-0.10742l5.2676-0.10742 5.1582 0.10742c0.93155 0.07176 2.2587 0.10742 3.9785 0.10742 0.21468 0 0.32227-0.21457 0.32227-0.64453 0-0.42996-0.10759-0.64453-0.32227-0.64453-3.0097 0-5.1236-0.71717-6.3418-2.1504-1.2182-1.5049-1.8281-4.0856-1.8281-7.7402v-49.229l1.8281 2.041 48.584 59.766c0.14355 0.14352 0.39571 0.17918 0.75391 0.10742 0.35817-0.07108 0.53711-0.17872 0.53711-0.32227v-57.076c0-3.6546 0.60793-6.2354 1.8262-7.7402 1.2899-1.5049 3.4753-2.2578 6.5566-2.2578 0.21543 0 0.32422-0.21457 0.32422-0.64453s-0.10879-0.64453-0.32422-0.64453l-4.1914 0.10547c-2.2931 0.14354-4.0494 0.2168-5.2676 0.2168-1.2899 0-3.0084-0.073252-5.1582-0.2168zm-359.66 0.00195c-0.14355 0-0.21484 0.21457-0.21484 0.64453s0.0713 0.64648 0.21484 0.64648c2.5798 0 4.8029 0.57215 6.666 1.7188 1.8632 1.1465 3.7248 3.1174 5.5879 5.9121l17.307 26.119-16.23 22.25c-2.0782 2.7947-4.2635 4.9463-6.5566 6.4512-2.2931 1.4332-4.6593 2.1484-7.0957 2.1484-0.21466 0-0.32032 0.21457-0.32032 0.64453 0 0.42996 0.10566 0.64648 0.32032 0.64648 1.7915 0 3.0828-0.03566 3.8711-0.10742l4.4062-0.10938 7.2012 0.10938c1.0749 0.07176 2.6881 0.10742 4.8379 0.10742 0.21465 0 0.32227-0.21653 0.32227-0.64648 0-0.42996-0.10762-0.64453-0.32227-0.64453-4.3712 0-6.5566-1.1821-6.5566-3.5469 0-1.2899 0.71529-2.9746 2.1484-5.0527l14.941-20.852 16.662 25.26c0.6446 0.93135 0.9668 1.6843 0.9668 2.2578 0 0.71678-0.46688 1.2175-1.3984 1.5039-0.93155 0.28641-2.4713 0.42969-4.6211 0.42969-0.14355 0-0.2168 0.21457-0.2168 0.64453 0 0.42996 0.0733 0.64453 0.2168 0.64453 2.2214 0 3.9758-0.03566 5.2656-0.10742l8.0625-0.10742 5.6973 0.10742c1.0749 0.07176 2.5431 0.10742 4.4062 0.10742 0.14355 0 0.21484-0.21457 0.21484-0.64453 0-0.42996-0.0713-0.64453-0.21484-0.64453-3.153 0-5.7318-0.53643-7.7383-1.6113-1.9348-1.1466-3.8342-3.1532-5.6973-6.0195l-17.844-26.766 15.049-20.529c4.7295-6.4494 9.5315-9.6758 14.404-9.6758 0.21463 1e-7 0.32031-0.21457 0.32031-0.64453s-0.10568-0.64453-0.32031-0.64453l-3.8711 0.10742c-1.7198 0.14354-3.188 0.21484-4.4062 0.21484-2.0782 0-4.4782-0.071299-7.2012-0.21484l-4.8379-0.10742c-0.2153-1e-7 -0.32226 0.21457-0.32226 0.64453s0.10696 0.64453 0.32226 0.64453c4.1562 0 6.2344 1.2198 6.2344 3.6562 0 1.4331-0.86025 3.404-2.5801 5.9121l-13.65 19.24-16.338-24.508c-0.6446-1.0031-0.9668-1.7561-0.9668-2.2578 0-0.71684 0.46493-1.2175 1.3965-1.5039 0.93162-0.35819 2.4732-0.53711 4.623-0.53711 0.21463 0 0.32032-0.21653 0.32032-0.64648-1e-5 -0.42996-0.10569-0.64453-0.32032-0.64453l-5.5898 0.10742c-3.4396 0.14354-6.3065 0.21484-8.5996 0.21484l-3.8691-0.10742c-1.8631-0.14354-3.226-0.21484-4.0859-0.21484zm84.326 0c-0.21465 0-0.32226 0.21457-0.32226 0.64453s0.10761 0.64453 0.32226 0.64453c2.7947 0 4.8371 0.21653 6.127 0.64648 1.3614 0.35832 2.2932 1.0378 2.7949 2.041 0.50173 1.0032 0.75195 2.5807 0.75195 4.7305v49.768c0 2.1498-0.25022 3.761-0.75195 4.8359-0.50166 1.0033-1.4334 1.6848-2.7949 2.043-1.2899 0.35819-3.3322 0.53711-6.127 0.53711-0.21465 0-0.32226 0.21457-0.32226 0.64453 0 0.42996 0.10761 0.64453 0.32226 0.64453 2.2215 0 3.9758-0.03566 5.2656-0.10742l7.2031-0.10742 7.5234 0.10742c1.2182 0.07176 2.9029 0.10742 5.0527 0.10742 0.14355 0 0.21485-0.21457 0.21485-0.64453 0-0.42996-0.0713-0.64453-0.21485-0.64453-2.8664 0-4.9445-0.17892-6.2344-0.53711-1.2897-0.35832-2.1858-1.0398-2.6875-2.043-0.42996-1.0033-0.64453-2.6145-0.64453-4.8359v-49.553c0-2.1498 0.21457-3.7272 0.64453-4.7305 0.5016-1.0749 1.3976-1.8259 2.6875-2.2559 1.3615-0.42996 3.4396-0.64648 6.2344-0.64648 0.14355 0 0.21485-0.21457 0.21485-0.64453s-0.0713-0.64453-0.21485-0.64453l-5.0527 0.10742c-3.0097 0.14354-5.517 0.21484-7.5234 0.21484-1.7915 0-4.1934-0.071299-7.2031-0.21484zm163.71 1.1816c6.306 0 11.823 1.4702 16.553 4.4082 4.8012 2.9381 8.4925 7.0586 11.072 12.361 2.6514 5.2312 3.9766 11.213 3.9766 17.949 0 6.3777-1.1821 11.861-3.5469 16.447-2.3648 4.5862-5.5892 8.061-9.6738 10.426-4.0845 2.3648-8.6699 3.5469-13.758 3.5469-6.0193 0-11.394-1.6113-16.123-4.8359-4.7295-3.2247-8.4209-7.5975-11.072-13.115-2.5798-5.5894-3.8691-11.715-3.8691-18.379 0-8.8858 2.3285-15.909 6.9863-21.068 4.6579-5.1595 11.143-7.7402 19.455-7.7402z" style={{fill:"#171717",whiteSpace:"pre"}}/>
-    </svg>
-)
-
-export default LexTitle
