--- Lexicon/client/src/components/branding/MwLogo.css
@@ -1,7 +1,0 @@
-.logo__footer--mw {
-    min-width: 60px;
-    min-height: 60px;
-    width: 2em;
-    height: 2em;
-    margin-top: 1em;
-}
