+++ Lexicon/client/src/views/auth/AuthView.js
@@ -2,8 +2,7 @@ import React, { useContext, useRef, useState } from 'react'
 import { useHistory } from 'react-router-dom'
 import { toast } from 'react-toastify'
 import { UserContext } from '../../providers/UserProvider'
-import LexLogo from '../../components/branding/LexLogo'
-import LexTitle from '../../components/branding/LexTitle'
+import { LexLogo, LexTitle } from '../../components/branding/Branding'
 import ChangeColorMode from '../../utils/ChangeColorMode'
 import './AuthView.css'
